Restructures routes for consistency
